First commit
A trivial implementation of the classic computer game minesweeper.

Runs in the terminal with ncurses. Squares can be navigated with wasd, hjkl, or arrow keys. Cells can be cleared with the spacebar or C, and cells can be flagged with f, ', or `. This implementation follows the traditional color scheme associated with quantities of adjacent mines.

There are still some elements of this code that look very C-like that could be cleaned up a bit. There's only so much we can do since ncurses is in C. But I definitely should totally rewrite and restructure the part of the program that parses arguments and creates Game objects.