Refactored out parse_args()
Argument parsing now uses the C-style getopt() interface, which is better than
what I was doing before. If we want to have a new game start when the user
finishes one game, we now have stack variables in main() that we can build
new games with. This argument parser is much cleaner too. Maybe we can make
it a switch statement though.